Rework detection if we need to reinitialize.
Until now we checked if the real PWM speed is the PWM speed set by us.
The problem with this logic is, that the boards firmware may not set the
speed we've given, but a slightly different value. For example my RX
580 sets 122 if given 127. This causes unnecessary reinitializations
which lead to fan speed spikes.

Work around this by checking the fan state instead of its speed. If
we've been reset to auto mode, switch back to manual mode an set the
appropriate speed. If we haven't been reset the normal logic will set
the correct speed.